#184: rendering optimization
