+++ src/timeline/components/timelineFilterControls.tsx
@@ -12,18 +12,8 @@ import {
 } from "../../workspace/redux/selectors";
 import { setFilters } from "../../workspace/redux/slice";
 import { IsDoneFilterValue } from "../../workspace/types/client";
-import { UseXZoomResult } from "../hooks/useXZoom";
-import { TimelineZoomControls } from "./timelineZoomControls";
 
-export interface TimelineFilterControlsProps {
-  zoom: UseXZoomResult;
-}
-
-export function TimelineFilterControls(
-  props: TimelineFilterControlsProps,
-): JSX.Element {
-  const { zoom } = props;
-
+function TimelineFilterControls(): JSX.Element {
   const dispatch = useRootDispatch();
   const filters = useSelector(filtersSelector);
   const schema = useSelector(schemaSelector);
@@ -83,9 +73,8 @@ export function TimelineFilterControls(
           );
         })}
       </Grid>
-      <Grid item xs={2}>
-        <TimelineZoomControls zoom={zoom} />
-      </Grid>
     </Grid>
   );
 }
+
+export default React.memo(TimelineFilterControls);
