+++ src/timeline/shapes/timelineField.tsx
@@ -1,8 +1,8 @@
 import React from "react";
-import { getEventRelativePosition } from "../../../visualization/functions";
-import { Point2D } from "../../../visualization/types";
-import { LabelingField, LabelingObject } from "../../../workspace/types/client";
-import { FieldBlock } from "../../types";
+import { getEventRelativePosition } from "../../visualization/functions";
+import { Point2D } from "../../visualization/types";
+import { LabelingField, LabelingObject } from "../../workspace/types/client";
+import { FieldBlock } from "../types";
 import TimelineFieldRect from "./timelineFieldRect";
 import TimelineRow from "./timelineRow";
