+++ src/timeline/shapes/timelineFieldRect.tsx
@@ -1,12 +1,12 @@
 import Konva from "konva";
 import React, { useRef } from "react";
 import { Rect } from "react-konva";
-import { getEventRelativePosition } from "../../../visualization/functions";
-import useMouseHover from "../../../visualization/hooks/useMouseHover";
-import { Point2D } from "../../../visualization/types";
-import { LabelingField } from "../../../workspace/types/client";
-import { darkerTimelineColors, lighterColors } from "../../timelineColors";
-import { FieldBlock } from "../../types";
+import { getEventRelativePosition } from "../../visualization/functions";
+import useMouseHover from "../../visualization/hooks/useMouseHover";
+import { Point2D } from "../../visualization/types";
+import { LabelingField } from "../../workspace/types/client";
+import { darkerTimelineColors, lighterColors } from "../timelineColors";
+import { FieldBlock } from "../types";
 
 export interface TimelineFieldRectProps {
   isSelected: boolean;
