+++ src/video/objects/polygonObjects.tsx
@@ -1,40 +1,42 @@
 import React from "react";
 import { Line } from "react-konva";
-import { LineBuilderStage } from "../../../editors/builders/lineBuilder";
-import { getFieldValues } from "../../../editors/functions";
+import { LineBuilderStage } from "../../editors/builders/lineBuilder";
+import { getFieldValues } from "../../editors/functions";
 import {
   FieldType,
   FinishedObjectProps,
   InProgressObjectProps,
   LabelingFieldAttributes,
   LabelingFieldValues,
-} from "../../../editors/types";
-import { HoverTooltip } from "../../../visualization/components/hoverTooltip";
-import { getLabelText } from "../../functions";
+} from "../../editors/types";
+import { HoverTooltip } from "../../visualization/components/hoverTooltip";
+import { getLabelText } from "../functions";
 import Sections, { SectionsShapeProps } from "../shapes/sections";
 
 export function getSectionsProps(
   values?: LabelingFieldValues,
   attributes?: LabelingFieldAttributes,
 ): SectionsShapeProps | null {
-  const line = values?.Line;
-  if (!line) return null;
-  const points = line[0].value;
+  const polygon = values?.Polygon;
+  if (!polygon) return null;
+  const points = polygon[0].value;
   if (!points) return null;
-  const stroke = attributes?.Line?.color;
-  return { points, stroke };
+  const stroke = attributes?.Polygon?.color;
+  return { points, stroke, closed: true, opacity: 0.9 };
 }
 
-export function LineInProgress(
+export function PolygonInProgress(
   props: InProgressObjectProps,
 ): JSX.Element | null {
   const { stage, value, fieldSchema } = props;
   if (stage <= LineBuilderStage.ONE_POINT) return null;
-  const lineProps = getSectionsProps(value, fieldSchema.attributes);
-  return lineProps && <Line {...lineProps} />;
+  const polygonProps = getSectionsProps(value, fieldSchema.attributes);
+  return polygonProps && <Line {...polygonProps} />;
 }
 
-export function LineFinished(props: FinishedObjectProps): JSX.Element | null {
+export function PolygonFinished(
+  props: FinishedObjectProps,
+): JSX.Element | null {
   const { frame, field, object, isSelected, onChange, onSelect } = props;
   const { fieldSchema } = field;
   const { perFrame, attributes } = fieldSchema;
@@ -43,13 +45,13 @@ export function LineFinished(props: FinishedObjectProps): JSX.Element | null {
     perFrame,
     frame,
   });
-  const lineProps = getSectionsProps(values, attributes);
+  const sectionsProps = getSectionsProps(values, attributes);
 
   return (
-    lineProps && (
+    sectionsProps && (
       <HoverTooltip text={getLabelText(object)}>
         <Sections
-          sectionsProps={lineProps}
+          sectionsProps={sectionsProps}
           object={object}
           onSelect={onSelect}
           isSelected={isSelected}
